--- components/postgresql/actions/execute-sql-query/execute-sql-query.mjs
@@ -1,28 +1,0 @@
-import postgresql from "../../postgresql.app.mjs";
-
-export default {
-  name: "Execute PostgreSQL Query",
-  version: "0.0.1",
-  key: "postgresql-execute-sql-query",
-  description: "Execute a custom PostgreSQL query. See [our docs](https://pipedream.com/docs/databases/working-with-sql) to learn more about working with SQL in Pipedream.",
-  type: "action",
-  props: {
-    postgresql,
-    // eslint-disable-next-line pipedream/props-description
-    sql: {
-      type: "sql",
-      auth: {
-        app: "postgresql",
-      },
-      label: "PostreSQL Query",
-    },
-  },
-  async run() {
-    const data = await this.postgresql.executeQuery({
-      text: this.sql.query,
-      values: this.sql.values,
-    });
-    console.table(data);
-    return data;
-  },
-};
